NewsScrapper/backend/requirements.txt
fastapi
sqlalchemy
beautifulsoup4
requests
google-generativeai>=0.8.0
openai>=1.3.0
python-dotenv
python-dateutil
lxml
feedparser
